requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
pyTelegramBotAPI==4.30.0
APScheduler==3.11.2
python-dotenv==1.2.1
